Import vendor/NetBSD/tests@b34f66deb
This change imports the latest snapshot of vendor/NetBSD/tests:
b34f66deb.

All code has been rebased or marked as an expected error to ensure that
the tests run cleanly on all Tier 1 architectures.

Any modifications that were made and were not upstreamed which were
specific to FreeBSD and were not easily mergeable have been removed and
placed in their own FreeBSD-specific tests to ease future merges.

Any and all FreeBSD-specific tests should be written as separate tests
expressed using either the ATF or GoogleTest frameworks.

+++ contrib/netbsd-tests/lib/libm/t_scalbn.c
@@ -247,7 +247,7 @@ ATF_TC_BODY(scalbnf_val, tc)
 		    "test %zu: return value %.8g instead of %.8g"
 		    " (error %.8g)",
 		    i, rv, tests[i].result,
-		    fabs((tests[i].result - rv)/tests[i].result));
+		    fabsf((tests[i].result - rv)/tests[i].result));
 	}
 }
